Limit the use of useTableColumnSort hook
