+++ frontend/.eslintrc
@@ -63,6 +63,14 @@
           {
             "group": ["~/api/**"],
             "message": "Read from '~/api' instead."
+          },
+          {
+            "group": ["~/components/table/**", "!~/components/table/useTableColumnSort"],
+            "message": "Read from '~/components/table' instead."
+          },
+          {
+            "group": ["~/components/table/useTableColumnSort"],
+            "message": "The data will be sorted in the table, don't use this hook outside of '~/components/table' repo. For more information, please check the props of the Table component."
           }
         ]
       }
