+++ frontend/src/components/table/index.ts
@@ -1,0 +1,8 @@
+export * from './types';
+export * from './const';
+
+export { default as Table } from './Table';
+export { default as useCheckboxTable } from './useCheckboxTable';
+
+export { default as TableRowTitleDescription } from './TableRowTitleDescription';
+export { default as CheckboxTd } from './CheckboxTd';
