+++ frontend/src/concepts/pipelines/content/tables/pipeline/PipelinesTable.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { TableVariant } from '@patternfly/react-table';
 import { PipelineKF } from '~/concepts/pipelines/kfTypes';
-import Table from '~/components/table/Table';
+import { Table } from '~/components/table';
 import PipelinesTableRow from '~/concepts/pipelines/content/tables/pipeline/PipelinesTableRow';
 import { FetchStateRefreshPromise } from '~/utilities/useFetchState';
 import { pipelineColumns } from '~/concepts/pipelines/content/tables/columns';
