+++ frontend/src/pages/projects/screens/detail/data-connections/DataConnectionsTable.tsx
@@ -1,6 +1,5 @@
 import * as React from 'react';
-import Table from '~/components/table/Table';
-
+import { Table } from '~/components/table';
 import { DataConnection } from '~/pages/projects/types';
 import { columns } from './data';
 import DataConnectionsTableRow from './DataConnectionsTableRow';
