+++ frontend/src/pages/projects/screens/projects/ProjectListView.tsx
@@ -1,8 +1,7 @@
 import * as React from 'react';
 import { Button, ButtonVariant, ToolbarItem } from '@patternfly/react-core';
 import { useNavigate } from 'react-router-dom';
-import Table from '~/components/table/Table';
-import useTableColumnSort from '~/components/table/useTableColumnSort';
+import { Table } from '~/components/table';
 import SearchField, { SearchType } from '~/pages/projects/components/SearchField';
 import { ProjectKind } from '~/k8sTypes';
 import { getProjectDisplayName, getProjectOwner } from '~/pages/projects/utils';
@@ -25,8 +24,7 @@ const ProjectListView: React.FC<ProjectListViewProps> = ({ allowCreate }) => {
   const navigate = useNavigate();
   const [searchType, setSearchType] = React.useState<SearchType>(SearchType.NAME);
   const [search, setSearch] = React.useState('');
-  const sort = useTableColumnSort<ProjectKind>(columns, 0);
-  const filteredProjects = sort.transformData(unfilteredProjects).filter((project) => {
+  const filteredProjects = unfilteredProjects.filter((project) => {
     if (!search) {
       return true;
     }
